% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_table.R
\name{save_table}
\alias{save_table}
\title{Save tables for this study}
\usage{
save_table(tab, nm, ...)
}
\arguments{
\item{tab}{The table}

\item{nm}{File name (without the extension)}

\item{...}{Extra parameters for \code{knitr::kable}}
}
\value{
Nothing. Just saves.
}
\description{
Save tables for this study
}
