% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_table.R
\name{save_table}
\alias{save_table}
\title{Save tables for this study}
\usage{
save_table(tab, nm, digits = 0, col_names = NULL)
}
\arguments{
\item{tab}{The table}

\item{nm}{File name (without the extension)}

\item{digits, col_names}{Parameters for `knitr::kable`}
}
\value{
Nothing. Just saves.
}
\description{
Save tables for this study
}
